Move entries from data models into separated and categorised modules
