+++ inference/core/interfaces/http/http_api.py
@@ -5,15 +5,54 @@ from typing import Any, List, Optional, Union
 import uvicorn
 from fastapi import Body, FastAPI, Path, Query, Request
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import FileResponse, JSONResponse, RedirectResponse, Response
+from fastapi.responses import JSONResponse, Response
 from fastapi.staticfiles import StaticFiles
 from fastapi_cprofile.profiler import CProfileMiddleware
 
-from inference.core import data_models as M
 from inference.core.devices.utils import GLOBAL_INFERENCE_SERVER_ID
+from inference.core.entities.requests.clip import (
+    ClipCompareRequest,
+    ClipImageEmbeddingRequest,
+    ClipTextEmbeddingRequest,
+)
+from inference.core.entities.requests.gaze import GazeDetectionInferenceRequest
+from inference.core.entities.requests.inference import (
+    ClassificationInferenceRequest,
+    InferenceRequest,
+    InferenceRequestImage,
+    InstanceSegmentationInferenceRequest,
+    ObjectDetectionInferenceRequest,
+)
+from inference.core.entities.requests.sam import (
+    SamEmbeddingRequest,
+    SamSegmentationRequest,
+)
+from inference.core.entities.requests.server_state import (
+    AddModelRequest,
+    ClearModelRequest,
+)
+from inference.core.entities.responses.clip import (
+    ClipCompareResponse,
+    ClipEmbeddingResponse,
+)
+from inference.core.entities.responses.gaze import GazeDetectionInferenceResponse
+from inference.core.entities.responses.inference import (
+    ClassificationInferenceResponse,
+    InferenceResponse,
+    InstanceSegmentationInferenceResponse,
+    MultiLabelClassificationInferenceResponse,
+    ObjectDetectionInferenceResponse,
+)
+from inference.core.entities.responses.sam import (
+    SamEmbeddingResponse,
+    SamSegmentationResponse,
+)
+from inference.core.entities.responses.server_state import (
+    ModelsDescriptions,
+    ServerVersionInfo,
+)
 from inference.core.env import (
     ALLOW_ORIGINS,
-    CLIP_MODEL_ID,
     CORE_MODEL_CLIP_ENABLED,
     CORE_MODEL_GAZE_ENABLED,
     CORE_MODEL_SAM_ENABLED,
@@ -202,15 +241,15 @@ class HttpInterface(BaseInterface):
         self.model_manager = model_manager
 
         def process_inference_request(
-            inference_request: M.InferenceRequest,
-        ) -> M.InferenceResponse:
+            inference_request: InferenceRequest,
+        ) -> InferenceResponse:
             """Processes an inference request by calling the appropriate model.
 
             Args:
-                inference_request (M.InferenceRequest): The request containing model ID and other inference details.
+                inference_request (InferenceRequest): The request containing model ID and other inference details.
 
             Returns:
-                M.InferenceResponse: The response containing the inference results.
+                InferenceResponse: The response containing the inference results.
             """
             self.model_manager.add_model(
                 inference_request.model_id, inference_request.api_key
@@ -220,14 +259,14 @@ class HttpInterface(BaseInterface):
             )
 
         def load_core_model(
-            inference_request: M.InferenceRequest,
+            inference_request: InferenceRequest,
             api_key: Optional[str] = None,
             core_model: str = None,
         ) -> None:
             """Loads a core model (e.g., "clip" or "sam") into the model manager.
 
             Args:
-                inference_request (M.InferenceRequest): The request containing version and other details.
+                inference_request (InferenceRequest): The request containing version and other details.
                 api_key (Optional[str]): The API key for the request.
                 core_model (str): The core model type, e.g., "clip" or "sam".
 
@@ -286,7 +325,7 @@ class HttpInterface(BaseInterface):
 
         @app.get(
             "/info",
-            response_model=M.ServerVersionInfo,
+            response_model=ServerVersionInfo,
             summary="Info",
             description="Get the server name and version number",
         )
@@ -294,9 +333,9 @@ class HttpInterface(BaseInterface):
             """Endpoint to get the server name and version number.
 
             Returns:
-                M.ServerVersionInfo: The server version information.
+                ServerVersionInfo: The server version information.
             """
-            return M.ServerVersionInfo(
+            return ServerVersionInfo(
                 name="Roboflow Inference Server",
                 version=__version__,
                 uuid=GLOBAL_INFERENCE_SERVER_ID,
@@ -307,7 +346,7 @@ class HttpInterface(BaseInterface):
 
             @app.get(
                 "/model/registry",
-                response_model=M.ModelsDescriptions,
+                response_model=ModelsDescriptions,
                 summary="Get model keys",
                 description="Get the ID of each loaded model",
             )
@@ -315,62 +354,62 @@ class HttpInterface(BaseInterface):
                 """Get the ID of each loaded model in the registry.
 
                 Returns:
-                    M.ModelsDescriptions: The object containing models descriptions
+                    ModelsDescriptions: The object containing models descriptions
                 """
                 models_descriptions = self.model_manager.describe_models()
-                return M.ModelsDescriptions.from_models_descriptions(
+                return ModelsDescriptions.from_models_descriptions(
                     models_descriptions=models_descriptions
                 )
 
             @app.post(
                 "/model/add",
-                response_model=M.ModelsDescriptions,
+                response_model=ModelsDescriptions,
                 summary="Load a model",
                 description="Load the model with the given model ID",
             )
             @with_route_exceptions
-            async def model_add(request: M.AddModelRequest):
+            async def model_add(request: AddModelRequest):
                 """Load the model with the given model ID into the model manager.
 
                 Args:
-                    request (M.AddModelRequest): The request containing the model ID and optional API key.
+                    request (AddModelRequest): The request containing the model ID and optional API key.
 
                 Returns:
-                    M.ModelsDescriptions: The object containing models descriptions
+                    ModelsDescriptions: The object containing models descriptions
                 """
 
                 self.model_manager.add_model(request.model_id, request.api_key)
                 models_descriptions = self.model_manager.describe_models()
-                return M.ModelsDescriptions.from_models_descriptions(
+                return ModelsDescriptions.from_models_descriptions(
                     models_descriptions=models_descriptions
                 )
 
             @app.post(
                 "/model/remove",
-                response_model=M.ModelsDescriptions,
+                response_model=ModelsDescriptions,
                 summary="Remove a model",
                 description="Remove the model with the given model ID",
             )
             @with_route_exceptions
-            async def model_remove(request: M.ClearModelRequest):
+            async def model_remove(request: ClearModelRequest):
                 """Remove the model with the given model ID from the model manager.
 
                 Args:
-                    request (M.ClearModelRequest): The request containing the model ID to be removed.
+                    request (ClearModelRequest): The request containing the model ID to be removed.
 
                 Returns:
-                    M.ModelsDescriptions: The object containing models descriptions
+                    ModelsDescriptions: The object containing models descriptions
                 """
 
                 self.model_manager.remove(request.model_id)
                 models_descriptions = self.model_manager.describe_models()
-                return M.ModelsDescriptions.from_models_descriptions(
+                return ModelsDescriptions.from_models_descriptions(
                     models_descriptions=models_descriptions
                 )
 
             @app.post(
                 "/model/clear",
-                response_model=M.ModelsDescriptions,
+                response_model=ModelsDescriptions,
                 summary="Remove all models",
                 description="Remove all loaded models",
             )
@@ -379,20 +418,20 @@ class HttpInterface(BaseInterface):
                 """Remove all loaded models from the model manager.
 
                 Returns:
-                    M.ModelsDescriptions: The object containing models descriptions
+                    ModelsDescriptions: The object containing models descriptions
                 """
 
                 self.model_manager.clear()
                 models_descriptions = self.model_manager.describe_models()
-                return M.ModelsDescriptions.from_models_descriptions(
+                return ModelsDescriptions.from_models_descriptions(
                     models_descriptions=models_descriptions
                 )
 
             @app.post(
                 "/infer/object_detection",
                 response_model=Union[
-                    M.ObjectDetectionInferenceResponse,
-                    List[M.ObjectDetectionInferenceResponse],
+                    ObjectDetectionInferenceResponse,
+                    List[ObjectDetectionInferenceResponse],
                 ],
                 summary="Object detection infer",
                 description="Run inference with the specified object detection model",
@@ -400,36 +439,36 @@ class HttpInterface(BaseInterface):
             )
             @with_route_exceptions
             async def infer_object_detection(
-                inference_request: M.ObjectDetectionInferenceRequest,
+                inference_request: ObjectDetectionInferenceRequest,
             ):
                 """Run inference with the specified object detection model.
 
                 Args:
-                    inference_request (M.ObjectDetectionInferenceRequest): The request containing the necessary details for object detection.
+                    inference_request (ObjectDetectionInferenceRequest): The request containing the necessary details for object detection.
 
                 Returns:
-                    Union[M.ObjectDetectionInferenceResponse, List[M.ObjectDetectionInferenceResponse]]: The response containing the inference results.
+                    Union[ObjectDetectionInferenceResponse, List[ObjectDetectionInferenceResponse]]: The response containing the inference results.
                 """
 
                 return process_inference_request(inference_request)
 
             @app.post(
                 "/infer/instance_segmentation",
-                response_model=M.InstanceSegmentationInferenceResponse,
+                response_model=InstanceSegmentationInferenceResponse,
                 summary="Instance segmentation infer",
                 description="Run inference with the specified instance segmentation model",
             )
             @with_route_exceptions
             async def infer_instance_segmentation(
-                inference_request: M.InstanceSegmentationInferenceRequest,
+                inference_request: InstanceSegmentationInferenceRequest,
             ):
                 """Run inference with the specified instance segmentation model.
 
                 Args:
-                    inference_request (M.InstanceSegmentationInferenceRequest): The request containing the necessary details for instance segmentation.
+                    inference_request (InstanceSegmentationInferenceRequest): The request containing the necessary details for instance segmentation.
 
                 Returns:
-                    M.InstanceSegmentationInferenceResponse: The response containing the inference results.
+                    InstanceSegmentationInferenceResponse: The response containing the inference results.
                 """
 
                 return process_inference_request(inference_request)
@@ -437,23 +476,23 @@ class HttpInterface(BaseInterface):
             @app.post(
                 "/infer/classification",
                 response_model=Union[
-                    M.ClassificationInferenceResponse,
-                    M.MultiLabelClassificationInferenceResponse,
+                    ClassificationInferenceResponse,
+                    MultiLabelClassificationInferenceResponse,
                 ],
                 summary="Classification infer",
                 description="Run inference with the specified classification model",
             )
             @with_route_exceptions
             async def infer_classification(
-                inference_request: M.ClassificationInferenceRequest,
+                inference_request: ClassificationInferenceRequest,
             ):
                 """Run inference with the specified classification model.
 
                 Args:
-                    inference_request (M.ClassificationInferenceRequest): The request containing the necessary details for classification.
+                    inference_request (ClassificationInferenceRequest): The request containing the necessary details for classification.
 
                 Returns:
-                    Union[M.ClassificationInferenceResponse, M.MultiLabelClassificationInferenceResponse]: The response containing the inference results.
+                    Union[ClassificationInferenceResponse, MultiLabelClassificationInferenceResponse]: The response containing the inference results.
                 """
 
                 return process_inference_request(inference_request)
@@ -463,13 +502,13 @@ class HttpInterface(BaseInterface):
 
                 @app.post(
                     "/clip/embed_image",
-                    response_model=M.ClipEmbeddingResponse,
+                    response_model=ClipEmbeddingResponse,
                     summary="CLIP Image Embeddings",
                     description="Run the Open AI CLIP model to embed image data.",
                 )
                 @with_route_exceptions
                 async def clip_embed_image(
-                    inference_request: M.ClipImageEmbeddingRequest,
+                    inference_request: ClipImageEmbeddingRequest,
                     api_key: Optional[str] = Query(
                         None,
                         description="Roboflow API Key that will be passed to the model during initialization for artifact retrieval",
@@ -480,12 +519,12 @@ class HttpInterface(BaseInterface):
                     Embeds image data using the OpenAI CLIP model.
 
                     Args:
-                        inference_request (M.ClipImageEmbeddingRequest): The request containing the image to be embedded.
+                        inference_request (ClipImageEmbeddingRequest): The request containing the image to be embedded.
                         api_key (Optional[str], default None): Roboflow API Key passed to the model during initialization for artifact retrieval.
                         request (Request, default Body()): The HTTP request.
 
                     Returns:
-                        M.ClipEmbeddingResponse: The response containing the embedded image.
+                        ClipEmbeddingResponse: The response containing the embedded image.
                     """
                     clip_model_id = load_clip_model(inference_request, api_key=api_key)
                     response = self.model_manager.infer_from_request(
@@ -500,13 +539,13 @@ class HttpInterface(BaseInterface):
 
                 @app.post(
                     "/clip/embed_text",
-                    response_model=M.ClipEmbeddingResponse,
+                    response_model=ClipEmbeddingResponse,
                     summary="CLIP Text Embeddings",
                     description="Run the Open AI CLIP model to embed text data.",
                 )
                 @with_route_exceptions
                 async def clip_embed_text(
-                    inference_request: M.ClipTextEmbeddingRequest,
+                    inference_request: ClipTextEmbeddingRequest,
                     api_key: Optional[str] = Query(
                         None,
                         description="Roboflow API Key that will be passed to the model during initialization for artifact retrieval",
@@ -517,12 +556,12 @@ class HttpInterface(BaseInterface):
                     Embeds text data using the OpenAI CLIP model.
 
                     Args:
-                        inference_request (M.ClipTextEmbeddingRequest): The request containing the text to be embedded.
+                        inference_request (ClipTextEmbeddingRequest): The request containing the text to be embedded.
                         api_key (Optional[str], default None): Roboflow API Key passed to the model during initialization for artifact retrieval.
                         request (Request, default Body()): The HTTP request.
 
                     Returns:
-                        M.ClipEmbeddingResponse: The response containing the embedded text.
+                        ClipEmbeddingResponse: The response containing the embedded text.
                     """
                     clip_model_id = load_clip_model(inference_request, api_key=api_key)
                     response = self.model_manager.infer_from_request(
@@ -537,13 +576,13 @@ class HttpInterface(BaseInterface):
 
                 @app.post(
                     "/clip/compare",
-                    response_model=M.ClipCompareResponse,
+                    response_model=ClipCompareResponse,
                     summary="CLIP Compare",
                     description="Run the Open AI CLIP model to compute similarity scores.",
                 )
                 @with_route_exceptions
                 async def clip_compare(
-                    inference_request: M.ClipCompareRequest,
+                    inference_request: ClipCompareRequest,
                     api_key: Optional[str] = Query(
                         None,
                         description="Roboflow API Key that will be passed to the model during initialization for artifact retrieval",
@@ -554,12 +593,12 @@ class HttpInterface(BaseInterface):
                     Computes similarity scores using the OpenAI CLIP model.
 
                     Args:
-                        inference_request (M.ClipCompareRequest): The request containing the data to be compared.
+                        inference_request (ClipCompareRequest): The request containing the data to be compared.
                         api_key (Optional[str], default None): Roboflow API Key passed to the model during initialization for artifact retrieval.
                         request (Request, default Body()): The HTTP request.
 
                     Returns:
-                        M.ClipCompareResponse: The response containing the similarity scores.
+                        ClipCompareResponse: The response containing the similarity scores.
                     """
                     clip_model_id = load_clip_model(inference_request, api_key=api_key)
                     response = self.model_manager.infer_from_request(
@@ -576,13 +615,13 @@ class HttpInterface(BaseInterface):
 
                 @app.post(
                     "/sam/embed_image",
-                    response_model=M.SamEmbeddingResponse,
+                    response_model=SamEmbeddingResponse,
                     summary="SAM Image Embeddings",
                     description="Run the Meta AI Segmant Anything Model to embed image data.",
                 )
                 @with_route_exceptions
                 async def sam_embed_image(
-                    inference_request: M.SamEmbeddingRequest,
+                    inference_request: SamEmbeddingRequest,
                     api_key: Optional[str] = Query(
                         None,
                         description="Roboflow API Key that will be passed to the model during initialization for artifact retrieval",
@@ -593,7 +632,7 @@ class HttpInterface(BaseInterface):
                     Embeds image data using the Meta AI Segmant Anything Model (SAM).
 
                     Args:
-                        inference_request (M.SamEmbeddingRequest): The request containing the image to be embedded.
+                        inference_request (SamEmbeddingRequest): The request containing the image to be embedded.
                         api_key (Optional[str], default None): Roboflow API Key passed to the model during initialization for artifact retrieval.
                         request (Request, default Body()): The HTTP request.
 
@@ -618,13 +657,13 @@ class HttpInterface(BaseInterface):
 
                 @app.post(
                     "/sam/segment_image",
-                    response_model=M.SamSegmentationResponse,
+                    response_model=SamSegmentationResponse,
                     summary="SAM Image Segmentation",
                     description="Run the Meta AI Segmant Anything Model to generate segmenations for image data.",
                 )
                 @with_route_exceptions
                 async def sam_segment_image(
-                    inference_request: M.SamSegmentationRequest,
+                    inference_request: SamSegmentationRequest,
                     api_key: Optional[str] = Query(
                         None,
                         description="Roboflow API Key that will be passed to the model during initialization for artifact retrieval",
@@ -635,7 +674,7 @@ class HttpInterface(BaseInterface):
                     Generates segmentations for image data using the Meta AI Segmant Anything Model (SAM).
 
                     Args:
-                        inference_request (M.SamSegmentationRequest): The request containing the image to be segmented.
+                        inference_request (SamSegmentationRequest): The request containing the image to be segmented.
                         api_key (Optional[str], default None): Roboflow API Key passed to the model during initialization for artifact retrieval.
                         request (Request, default Body()): The HTTP request.
 
@@ -662,13 +701,13 @@ class HttpInterface(BaseInterface):
 
                     @app.post(
                         "/gaze/gaze_detection",
-                        response_model=List[M.GazeDetectionInferenceResponse],
+                        response_model=List[GazeDetectionInferenceResponse],
                         summary="Gaze Detection",
                         description="Run the gaze detection model to detect gaze.",
                     )
                     @with_route_exceptions
                     async def gaze_detection(
-                        inference_request: M.GazeDetectionInferenceRequest,
+                        inference_request: GazeDetectionInferenceRequest,
                         api_key: Optional[str] = Query(
                             None,
                             description="Roboflow API Key that will be passed to the model during initialization for artifact retrieval",
@@ -705,10 +744,10 @@ class HttpInterface(BaseInterface):
                 "/{dataset_id}/{version_id}",
                 # Order matters in this response model Union. It will use the first matching model. For example, Object Detection Inference Response is a subset of Instance segmentation inference response, so instance segmentation must come first in order for the matching logic to work.
                 response_model=Union[
-                    M.InstanceSegmentationInferenceResponse,
-                    M.ObjectDetectionInferenceResponse,
-                    M.ClassificationInferenceResponse,
-                    M.MultiLabelClassificationInferenceResponse,
+                    InstanceSegmentationInferenceResponse,
+                    ObjectDetectionInferenceResponse,
+                    ClassificationInferenceResponse,
+                    MultiLabelClassificationInferenceResponse,
                     Any,
                 ],
                 response_model_exclude_none=True,
@@ -799,7 +838,7 @@ class HttpInterface(BaseInterface):
                     # Other parameters described in the function signature...
 
                 Returns:
-                    Union[M.InstanceSegmentationInferenceResponse, M.ObjectDetectionInferenceResponse, M.ClassificationInferenceResponse, M.MultiLabelClassificationInferenceResponse, Any]: The response containing the inference results.
+                    Union[InstanceSegmentationInferenceResponse, ObjectDetectionInferenceResponse, ClassificationInferenceResponse, MultiLabelClassificationInferenceResponse, Any]: The response containing the inference results.
                 """
                 model_id = f"{dataset_id}/{version_id}"
 
@@ -812,7 +851,7 @@ class HttpInterface(BaseInterface):
                     overlap /= 100
 
                 if image is not None:
-                    request_image = M.InferenceRequestImage(type="url", value=image)
+                    request_image = InferenceRequestImage(type="url", value=image)
                 else:
                     if "Content-Type" not in request.headers:
                         raise ContentTypeMissing(
@@ -821,7 +860,7 @@ class HttpInterface(BaseInterface):
                     if "multipart/form-data" in request.headers["Content-Type"]:
                         form_data = await request.form()
                         base64_image_str = form_data["file"].file
-                        request_image = M.InferenceRequestImage(
+                        request_image = InferenceRequestImage(
                             type="multipart", value=base64_image_str
                         )
                     elif (
@@ -830,7 +869,7 @@ class HttpInterface(BaseInterface):
                         or "application/json" in request.headers["Content-Type"]
                     ):
                         data = await request.body()
-                        request_image = M.InferenceRequestImage(
+                        request_image = InferenceRequestImage(
                             type=image_type, value=data
                         )
                     else:
@@ -863,16 +902,16 @@ class HttpInterface(BaseInterface):
                 self.model_manager.add_model(request_model_id, api_key)
 
                 task_type = self.model_manager.get_task_type(request_model_id)
-                inference_request_type = M.ObjectDetectionInferenceRequest
+                inference_request_type = ObjectDetectionInferenceRequest
                 args = dict()
                 if task_type == "instance-segmentation":
-                    inference_request_type = M.InstanceSegmentationInferenceRequest
+                    inference_request_type = InstanceSegmentationInferenceRequest
                     args = {
                         "mask_decode_mode": mask_decode_mode,
                         "tradeoff_factor": tradeoff_factor,
                     }
                 elif task_type == "classification":
-                    inference_request_type = M.ClassificationInferenceRequest
+                    inference_request_type = ClassificationInferenceRequest
 
                 inference_request = inference_request_type(
                     model_id=request_model_id,
