+++ inference/core/interfaces/udp/udp_stream.py
@@ -8,7 +8,7 @@ from typing import Union
 import supervision as sv
 from PIL import Image
 
-from inference.core import data_models as M
+import inference.core.entities.requests.inference
 from inference.core.env import (
     API_KEY,
     CLASS_AGNOSTIC_NMS,
@@ -103,7 +103,9 @@ class UdpStream(BaseInterface):
         self.ip_broadcast_port = ip_broadcast_port
         self.json_response = json_response
 
-        self.inference_request_type = M.ObjectDetectionInferenceRequest
+        self.inference_request_type = (
+            inference.core.entities.requests.inference.ObjectDetectionInferenceRequest
+        )
 
         self.UDPServerSocket = socket.socket(
             family=socket.AF_INET, type=socket.SOCK_DGRAM
