+++ inference/core/managers/decorators/base.py
@@ -1,4 +1,5 @@
-from inference.core.data_models import InferenceRequest, InferenceResponse
+from inference.core.entities.requests.inference import InferenceRequest
+from inference.core.entities.responses.inference import InferenceResponse
 from inference.core.managers.base import Model, ModelManager
