+++ inference/core/managers/decorators/fixed_size_cache.py
@@ -1,7 +1,8 @@
 from collections import deque
 from typing import List
 
-from inference.core.data_models import InferenceRequest, InferenceResponse
+from inference.core.entities.requests.inference import InferenceRequest
+from inference.core.entities.responses.inference import InferenceResponse
 from inference.core.managers.base import Model, ModelManager
 from inference.core.managers.decorators.base import ModelManagerDecorator
 from inference.core.managers.entities import ModelDescription
