+++ inference/core/models/base.py
@@ -3,7 +3,8 @@ from typing import Any, List, Tuple, Union
 
 import numpy as np
 
-from inference.core.data_models import InferenceRequest, InferenceResponse
+from inference.core.entities.requests.inference import InferenceRequest
+from inference.core.entities.responses.inference import InferenceResponse
 from inference.core.models.types import PreprocessReturnMetadata
