+++ inference/core/models/instance_segmentation_base.py
@@ -3,7 +3,7 @@ from typing import Any, List, Tuple, Union
 
 import numpy as np
 
-from inference.core.data_models import (
+from inference.core.entities.responses.inference import (
     InferenceResponseImage,
     InstanceSegmentationInferenceResponse,
     InstanceSegmentationPrediction,
