+++ inference/models/sam/segment_anything.py
@@ -10,12 +10,14 @@ import torch
 from segment_anything import SamPredictor, sam_model_registry
 from shapely.geometry import Polygon as ShapelyPolygon
 
-from inference.core.data_models import (
-    InferenceRequestImage,
+from inference.core.entities.requests.inference import InferenceRequestImage
+from inference.core.entities.requests.sam import (
     SamEmbeddingRequest,
-    SamEmbeddingResponse,
     SamInferenceRequest,
     SamSegmentationRequest,
+)
+from inference.core.entities.responses.sam import (
+    SamEmbeddingResponse,
     SamSegmentationResponse,
 )
 from inference.core.env import SAM_MAX_EMBEDDING_CACHE_SIZE, SAM_VERSION_ID
